transport: mark bundle fetch unused
This is actually not ideal, because it means we fetch more
objects from the bundle than we actually need. But the
bundle does not have any way of limiting the set of objects
based on reachability. And we have to fetch it as a whole
anyway, so it's just copying objects locally.

Signed-off-by: Jeff King <peff@peff.net>